+++ package.json
@@ -23,7 +23,7 @@
     "@radix-ui/react-tooltip": "^1.2.7",
     "@react-spring/three": "^9.7.5",
     "@react-three/drei": "^9.122.0",
-    "@react-three/fiber": "^8.18.0",
+    "@react-three/fiber": "^9.3.0",
     "@shadergradient/react": "^2.1.0",
     "@tailwindcss/typography": "^0.5.16",
     "@types/react-dom": "18.0.11",
